Amélioration: fiche de carac: aura de feu

+++ fonctions_sql/f_perso_aura_feu.sql
@@ -12,7 +12,7 @@ declare
 
   personnage alias for $1;
   code_retour numeric;
-  v_aura_feu integer;
+  v_aura_feu numeric;
 
 begin
   code_retour := 0;
